fixed sign in, security optimization and db integration for users and score.
sign in and login pages have been created as well as validation if used incorrectly.
created a high score section in program where integration from database is currently being implemented.
implemented a page where the user is indicated their score, time and the list of songs that where utilised during the quiz as well as stating that spotify is the producers of the images as well as songs previews.
security optimizations